Add skeleton of starboard cog; improve menu handling
We were previously not escaping mentions right in `qotd list` and
`qotd suggest`. I think everyone and here could have maybe got
through? Not sure about the defaults there.

Also improved the handling in Markov where previously we'd been
1. allowing mentions but 2. editing the message from a dummy one so
they didn't ping, which I think was out of some idea of preserving the
exact text of an exclusion but it seems inelegant.