markov: avoid float division in uint_to_bytes
It seems as though the float division could cause an issue where the
actual value should be an integer but the float representation is
slightly higher than that, causing the addition of an extra unnecessary
byte. Although this works fine as long as the same function is used,
it could cause problems if seemingly inconsequential details change.

I switch to using divmod and checking for the presence of a remainder.
This should ensure use of the exact minimum required number of bytes in
the blob.